moved ByteArrayReadableSeekableByteChannel to src main o.a.a.vector.util
